% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{deleteUI}
\alias{deleteUI}
\title{按纽生成生成器界面}
\usage{
deleteUI(
  tabTitle = "清除数据",
  colTitles = c("操作区域", "操作区域", "显示区域"),
  widthRates = c(4, 8, 12),
  func_left = deleteUI_left,
  func_right = deleteUI_right,
  func_bottom = deleteUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
按纽生成生成器界面
}
\examples{
deleteUI()
}
